+++ frontend/src/views/Dash.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+function Dash() {
+    return (
+        <>
+        <h1>Dashboard.</h1>
+        <div>Dash</div>
+        </>
+    )
+}
+
+export default Dash